refactor(clase01/ejercicio02): usar int en la duración y renombrar a titulo
La duración se expresa en segundos, así que int alcanza y evita el error
de redondeo que acumulaba duracionTotal al sumar y restar doubles.

Cancion.nombre pasa a titulo para coincidir con el enunciado.

+++ src/ar/edu/unlu/poo/clase01/ejercicio02_playlist/Playlist.java
@@ -4,7 +4,7 @@ public class Playlist {
     private NodoCancion primeraCancion;
     private NodoCancion ultimaCancion;
     private int numeroCanciones;
-    private double duracionTotal;
+    private int duracionTotal;
     private final String nombre;
 
     public Playlist(String nombre) {
@@ -54,7 +54,7 @@ public class Playlist {
         return numeroCanciones == 0;
     }
 
-    public double obtenerDuracionTotal() {
+    public int obtenerDuracionTotal() {
         return duracionTotal;
     }
